test/requirements.txt
requests>=2.11.1
selenium>=3.14.0
lxml>=4.0.0
public>=2018.11.20
pycurl>=7.19.0
virtualenv>=12.0
mitmproxy>=4.0.0
ldaptor>=16.0
psutil>=3.0.0
public>=0.0.0

# For running tests on helper
Babel>=2.6.0
Flask-Login>=0.4.0
Flask-Babel>=1.0.0
sqlalchemy>=1.3.10
iso-639>=0.4.5
unidecode>=1.0.0
Flask-Principal>=0.3.2

# Gdrive integration
fs.googledrivefs>=1.7.0